requirements.txt
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dynaconf==3.2.3
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
outcome==1.2.0
packaging==23.2
pandas==2.1.1
peewee==3.17.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
webdriver-manager==4.0.1
Werkzeug==3.0.0
wget==3.2
wsproto==1.2.0
